Refactored matrix targeting out of MatrixFlow
